:recycle: move slack verifier to lib/slack
